Simplify snake entity internals
The idea for tracking the body of the snake as a sequence of vectors
seemed like a good idea, but it ended up being too complicated when
working to implement ending the game when the user crashed the snake.
This refactors the internals to simply track itself using a sequence of
cells. I also changed how the snake was presented on the screen to be a
single letter, 'X'. This clarifies which "cell" on the screen is
actually occupied by the snake.